Moved `InMemoryFeedStore` to its own file
